refactor(ai-service): Step 11 — relocate ML services + delete app/services/
The five ML-named service files relocate into their target bands.
Internal file splits per Page 5 of the diagram are deferred to a
follow-up PR; this commit does pure relocations only.

Moves:
  services/tire_grip_analysis_service.py
    → app/features/tire_grip.py
      (NOT app/ml/ — physics math, no fit / predict)
  services/transformer_model.py
    → app/ml/transformer/model.py
  services/imitate_expert_learning_service.py
    → app/ml/imitation/service.py
  services/segment_classifier_service.py
    → app/ml/segment_classifier/service.py
  services/full_dataset_ml_service.py
    → app/pipelines/training/full_dataset.py
      (it is an orchestrator, not a model — belongs with pipelines)
  services/segment_updater.py
    → app/pipelines/inference/segment_updater.py
  services/telemetry_segment_visualizer.py
    → app/pipelines/inference/visualizer.py

The whole app/services/ directory is deleted. The previously-archived
scikit_imitation_learning_service.py moves to app/_archive/ instead
of being lost.

Importers updated (9 files across api/, ui/, scripts/).

One subtle issue surfaced and fixed: pipelines.chat.__init__ (AIService)
imports Full_dataset_TelemetryMLService from pipelines.training.full_dataset,
and full_dataset imported TelemetryLLMOrchestrator from
pipelines.chat.orchestrator at module load. Since orchestrator is a
SIBLING of chat/__init__, loading it required chat/__init__ to finish
first — circular. Pre-refactor these two files lived in different
packages (services/ai_service vs services/llm/telemetry_llm_orchestrator)
so no cycle. Fix: defer the orchestrator import to inside the
Full_dataset_TelemetryMLService.__init__ method.

Validated:
  • all 7 new module paths import cleanly
  • AIService() instantiates end-to-end; its inner orchestrator is
    a real TelemetryLLMOrchestrator (lazy import resolved)
  • /health 200
  • app/services/ no longer exists

After this commit, app/services/ is gone. What's left to clean up:
  • Internal file splits of the relocated ML services (Page 5 detail)
  • app/core/ + app/config/ → app/infra/config/ (Step 12)
  • app/analyzers/ deletion (empty since the original code)
  • Add POST /annotation/run endpoint (Step 13)
