--- acla_ai_service/app/services/__init__.py
@@ -1,20 +1,0 @@
-"""
-Services module initialization
-"""
-
-from app.pipelines.chat import AIService
-from app.integrations.backend.client import BackendService
-from .tire_grip_analysis_service import TireGripAnalysisService, tire_grip_analysis_service
-from .segment_classifier_service import SegmentClassifierService, segment_classifier
-
-__all__ = [
-    "AIService",
-    "BackendService",
-    "AnalysisService",
-    "TireGripAnalysisService", 
-    "tire_grip_analysis_service",
-    "TelemetryPromptDatasetBuilder",
-    "PromptBuilderConfig",
-    "SegmentClassifierService",
-    "segment_classifier",
-]
