+++ acla_ai_service/scripts/run_full_pipeline.py
@@ -14,9 +14,9 @@ faulthandler.enable()
 # parents[1] = acla_ai_service (or /app in docker)
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 
-from app.services.full_dataset_ml_service import Full_dataset_TelemetryMLService
+from app.pipelines.training.full_dataset import Full_dataset_TelemetryMLService
 from app.config.pipeline_config import PipelineConfig
-from app.services.segment_classifier_service import segment_classifier
+from app.ml.segment_classifier.service import segment_classifier
 
 logger = logging.getLogger("run_full_pipeline")
